Implement changes for Multistep form example, document in README
